Do not elide amounts in generated reports
It seems that, since [`hledger 1.19`][hledger_commit], report lines with
multiple commodities are elided by default. This is the difference
between having a total row displaying `USD 1,000.00, 2 more..`, and
having the totals for all the different commodities displayed in the
generated report.

With this change, reports now display all commodities.

[hledger_commit]: https://github.com/simonmichael/hledger/commit/2739a70a38b28114d9ab042cdf3597f53337dc02